feat: restructure post-related files
